+++ Packages/com.chark.game-management/CHANGELOG.md
@@ -16,6 +16,7 @@
 - Caching to `DefaultEntityManager`.
 - `OnStarted` to `GameManager` which can be used to hook into the `Start()` method.
 - Very biased logging API which formats log messages in a consistent format.
+- `async` message support.
 
 ### Changed